chore: Move `zephyr/include/net.h` to drop the `include` subdirectory
It's a header file only used by source file `net.c` which it can
live next to.

Ticket: MEN-7692
Changelog: none
